+++ Pronto-MIA/Pronto-MIA/DataAccess/Managers/FileManager.cs
@@ -3,9 +3,11 @@ namespace Pronto_MIA.DataAccess.Managers
     using System;
     using System.IO;
     using System.Threading.Tasks;
+    using HotChocolate.Execution;
     using HotChocolate.Types;
     using Microsoft.Extensions.Configuration;
     using Microsoft.Extensions.Logging;
+    using Pronto_MIA.BusinessLogic.API.EntityExtensions;
 
     /// <summary>
     /// Class responsible for the lifecycle of a user within the application.
@@ -39,14 +41,14 @@ namespace Pronto_MIA.DataAccess.Managers
         /// <param name="file">The file from which the extension should be
         /// extracted.</param>
         /// <returns>The file extension.</returns>
-        /// <exception cref="ArgumentException">If the file has no file
-        /// extension or the filename contains invalid characters.</exception>
+        /// <exception cref="QueryException">Throws if the file that was
+        /// uploaded has no file extension.</exception>
         public static string GetFileExtension(IFile file)
         {
             var ext = Path.GetExtension(file.Name);
             if (string.IsNullOrEmpty(ext))
             {
-                throw new ArgumentException("No file extension found");
+                throw DataAccess.Error.FileOperationError.AsQueryException();
             }
 
             return ext!;
@@ -61,21 +63,35 @@ namespace Pronto_MIA.DataAccess.Managers
         /// <param name="fileName">The name the file should be saved as.</param>
         /// <param name="file">The file that will be saved.</param>
         /// <returns>Empty task.</returns>
+        /// <exception cref="QueryException">Throws if an error occured during
+        /// file creation.</exception>
         public async Task
             Create(string subDirectory, string fileName, IFile file)
         {
             var ext = FileManager.GetFileExtension(file);
 
-            Directory.CreateDirectory(
-                this.GetDirectoryPath(subDirectory));
-            await using Stream stream = file.OpenReadStream();
-            await using FileStream fs = File.Create(
-                this.GetFilePath(subDirectory, fileName, ext));
-            await stream.CopyToAsync(fs);
-            stream.Close();
-            fs.Close();
-            this.logger.LogDebug(
-                "File {File} has been created", fileName);
+            try
+            {
+                Directory.CreateDirectory(
+                    this.GetDirectoryPath(subDirectory));
+                await using Stream stream = file.OpenReadStream();
+                await using FileStream fs = File.Create(
+                    this.GetFilePath(subDirectory, fileName, ext));
+                await stream.CopyToAsync(fs);
+                stream.Close();
+                fs.Close();
+                this.logger.LogDebug(
+                    "File {File} has been created",
+                    subDirectory + "/" + fileName + "." + ext);
+            }
+            catch (Exception error)
+            {
+                this.logger.LogError(
+                    "File {File} could not be created. Error: \n {error}",
+                    subDirectory + "/" + fileName + "." + ext,
+                    error);
+                throw DataAccess.Error.FileOperationError.AsQueryException();
+            }
         }
 
         /// <summary>
@@ -86,18 +102,27 @@ namespace Pronto_MIA.DataAccess.Managers
         /// be found.</param>
         /// <param name="fileName">The name of the file.</param>
         /// <param name="fileExtension">The file extension of the file.</param>
-        /// <returns>Error if something went wrong or null if successful.
-        /// </returns>
-        public DataAccess.Error? Remove(
+        /// <exception cref="QueryException">Throws if the given file could not
+        /// be removed.</exception>
+        public void Remove(
             string subDirectory,
             string fileName,
             string fileExtension)
         {
-            File.Delete(
-                this.GetFilePath(subDirectory, fileName, fileExtension));
-            this.logger.LogDebug(
-                "File {File} has been removed", fileName);
-            return null;
+            try
+            {
+                File.Delete(
+                    this.GetFilePath(subDirectory, fileName, fileExtension));
+                this.logger.LogDebug(
+                    "File {File} has been removed", fileName);
+            }
+            catch (Exception error)
+            {
+                this.logger.LogError(
+                    "File {File} could not be removed. Error: \n {error}",
+                    subDirectory + "/" + fileName + "." + fileExtension,
+                    error);
+            }
         }
 
         /// <summary>