+++ Pronto-MIA/Pronto-MIA/Program.cs
@@ -3,6 +3,8 @@ namespace Pronto_MIA
     using Microsoft.AspNetCore.Hosting;
     using Microsoft.Extensions.Configuration;
     using Microsoft.Extensions.Hosting;
+    using Microsoft.Extensions.Logging;
+    using Microsoft.Extensions.Logging.Console;
 
     /// <summary>
     /// Class for starting the application.