+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Mutation endpoint "RegisterFcmToken" to register a fcm device token.
+- Mutation endpoint "UnregisterFcmToken" to remove a fcm device token.
+- Mutation endpoint "UpdateDeploymentPlan" to adjust a deployment plan.
+- Mutation endpoint "RemoveDeploymentPlan" to remove a deployment plan.
+
+### Changed
+- Improved logging.
+- Improved error handling.
+- Errors now return a traceId in order to correlate the error with the server.
 
 ## [0.1.0] - 2021-04-06
 ### Added